requirements.txt
# Core
django==5.2.6
str2bool==1.1

# UI & Admin UI
django-jazzmin==3.0.1
django-admin-black==1.0.19

# Tools
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
django-filter>=24.0  # For DRF filtering
requests==2.32.3
pandas>=2.2.3
graphviz==0.20.3
astor==0.8.1 

# AI
anthropic==0.34.2
google-genai  # Google Gemini

# Deployment
whitenoise==6.7.0
gunicorn==23.0.0

# DB
#psycopg2-binary==2.9.9
#mysqlclient==2.1.1
django-dbbackup==4.2.1
mssql-django==1.6
pyodbc>=5.2.0

# Database ORM (optional, for advanced users)
sqlalchemy>=1.4.0

# Data processing and analysis
numpy>=1.21.0
openpyxl>=3.0.9

# Data validation and utilities
pydantic>=1.8.0
python-dateutil>=2.8.0

# Reporting and output
tabulate>=0.8.0
reportlab>=3.6.0
jinja2>=3.0.0

# CLI and utilities
click>=8.0.0
colorama>=0.4.0
python-dotenv>=0.19.0

